Billing and Dynamic Updates

+++ New Intellij/Utilities/RandomSeatsFiller.java
@@ -12,7 +12,7 @@ public class RandomSeatsFiller
     private static final int MIN_SEATS_TO_BE_FILLED = 40;
     private static final int MAX_SEATS_THAT_CAN_BE_FILLED = 70;
     private static final int MAX_SEATS_THAT_CAN_BE_FILLED_APPEND = 5;
-    private static final int MAX_FILLED_PERCENTAGE = 50;
+    private static final float MAX_FILLED_PERCENTAGE = 50;
     private static final int MAX_CONSECUTIVE_SEATS = 8;
 
     public static void Fill(Movie movie, CinemaHall hall, Date date, int time, Path path, int totalSeats, boolean isAppend)
@@ -28,7 +28,7 @@ public class RandomSeatsFiller
             while (seatsFilled <= seatsToBeFilled)
             {
                 if (BookingManager.GetBookedSeats(movie, hall, date, time).size() >=
-                        (totalSeats * MAX_FILLED_PERCENTAGE / 100)) return;
+                        (totalSeats * MAX_FILLED_PERCENTAGE / 100f)) return;
 
                 int randSeat = randomGenerator.nextInt(1, totalSeats + 1);
                 int consecutiveSeats = randomGenerator.nextInt(0, MAX_CONSECUTIVE_SEATS);
